+++ src/univix_util.cpp
@@ -1,7 +1,7 @@
 #include "univix_util.h"
 
-int util::bit_count[256];
-int util::bit_count2x[256];
+int16_t util::bit_count[256];
+int16_t util::bit_count2x[256];
 byte util::first_bit_mask[256];
 byte util::first_bit_offset[256];
 uint32_t util::left_mask32[32];
